follow-up on 4b657747 and move the public fixtures into `kotti.tests` itself
